+++ frontend/src/components/project/TaskCreateDetail/Calendar.jsx
@@ -1,7 +1,7 @@
-function CalendarDetail() {
+const Calendar = () => {
     return (
         <div>캘린더창입니다</div>
     )
 }
 
-export default CalendarDetail
+export default Calendar
