+++ frontend/src/components/project/TaskCreateDetail/Drawer.jsx
@@ -1,7 +1,7 @@
-function DrawerDetail() {
+const Drawer = () => {
     return (
         <div>서랍선택 창입니다</div>
     )
 }
 
-export default DrawerDetail
+export default Drawer
