+++ frontend/src/components/project/TaskCreateDetail/Memo.jsx
@@ -1,4 +1,4 @@
-function Memo() {
+const Memo = () => {
     return (
         <div>Memo 창입니다</div>
     )
