+++ frontend/src/components/project/TaskCreateDetail/Priority.jsx
@@ -1,7 +1,7 @@
-function PriorityDetail() {
+const Priority = () => {
     return (
         <div>서랍선택 창입니다</div>
     )
 }
 
-export default PriorityDetail
+export default Priority
