+++ frontend/src/components/project/TaskCreateDetail/Reminder.jsx
@@ -1,7 +1,7 @@
-function ReminderDetail() {
+const Reminder = () => {
     return (
         <div>Reminder 창입니다</div>
     )
 }
 
-export default ReminderDetail
+export default Reminder
